Some refactoring into library functions and types
